Rename build_contract_interface to contract_interface_builder
