+++ src/vm/analysis/contract_interface_builder/mod.rs
